requirements.txt
requests
python-dotenv
fastapi
uvicorn
pytest
httpx
numpy
pandas
matplotlib
seaborn
scikit-learn
pyarrow
xgboost
tabulate
python-dotenv
requests
pydantic
fastapi
uvicorn
python-jose[cryptography]
python-multipart
pytest
httpx